add prototypes by url
this updates prototype management to be able to add a prototype by url.
A url pointing to a jason prototype structure. for example, put
prototype on github and supply a url to add prototype will save you a
lot of typing.

This also adds a list of prototypes to add at creation. currently only
adding by a prototype by url is supported at creation. In prototype
structure add a list of prototypes add attributes (see documentation
for prototypes), they will be added at creation.