add recursion for prototype_subscriptions
I have added recursive functions for dynamically finding a pico eci
from a path. this pico must be a child of your owner pico. just provide
a hash path starting at “owner” using “__Oedipus_” as the dynamically
created child in the tree path.

if you start sending events to children during the initial set up of
the pico system some things may fail to finish. This behavior has been
seen mostly with subscriptions and with channels. This could be a
result of extra load on the server causing operation time outs.